requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
pandas==2.2.3
openpyxl==3.1.5
anthropic==0.40.0
python-dotenv==1.0.1
pyxlsb==1.0.10
